% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_classification}
\alias{check_classification}
\title{Evaluate sequences classification quality.
This function is used to check the efficiency of a tree-based sequence
  classification given a known lineage string.}
\usage{
check_classification(predicted, actual)
}
\arguments{
\item{predicted, actual}{semicolon-delimited lineage strings, where the former
is predicted by a classification tree and the latter is the true lineage.}
}
\value{
A signed integer.
  A negative number indicates an underclassification (a correct subset of the actual
  lineage), a zero represents a full correct classification,
  and a positive number indicates a over- or mis-classification.
}
\description{
Evaluate sequences classification quality.
This function is used to check the efficiency of a tree-based sequence
  classification given a known lineage string.
}
