% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_metadata}
\alias{check_metadata}
\title{Identify sequences with potentially incorrect lineage metadata}
\usage{
check_metadata(x, threshold = 5)
}
\arguments{
\item{x}{a DNAbin object with lineage attributes}

\item{threshold}{numeric. The minimum sequence weight to
prompt addition to the list of potentially incorrect
sequences. Defaults to 5.}
}
\value{
a character vector giving the names of the potentially incorrect sequences
}
\description{
This function evaluates a list of DNA barcode
   sequences (a "DNAbin" object with lineage attributes)
   and returns the names of the sequences that
   may require further checking before being used as training
   data for downstream tree-learning operations.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
