% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Taxonomic classification for DNA barcode sequences.}
\usage{
classify(x, tree, threshold = 0.9, decay = TRUE, ping = TRUE, cores = 1,
  scores = TRUE, paths = FALSE, seqlengths = FALSE, threshs = FALSE,
  minscores = FALSE, minlengths = FALSE, maxlengths = FALSE)
}
\arguments{
\item{x}{a \code{"DNAbin"} object, either as a vector or a list of vectors.}

\item{tree}{an object of class \code{"insect"}.}

\item{threshold}{numeric value between 0 and 1 giving the minimum
Akaike weight for the recursive classification procedure
to continue toward the leaves of the tree.}

\item{decay}{logical indicating whether the decision to terminate the
classification process should occur based on decaying Akaike weights
(at each node, the Akaike weight of the selected model is multiplied by
the Akaike weight of the selected model at the parent node) or whether
each Akaike weight should be calculated independently of that of the
parent node. Defaults to the latter (FALSE).}

\item{ping}{logical indicating whether an exact-match search should
be carried out before applying the classification algorithm.
If TRUE (the default value) and the query sequence is identical to
at least one of the training sequences used to learn the tree,
the common ancestor of the matching training sequences is returned
as a semicolon-delimited lineage string with an associated score value of 1.
The lineage string will generally specify the taxonomic ID to species level
but may be to genus/family, etc if the barcoding marker lacks resolution.}

\item{cores}{integer giving the number of CPUs to parallelize the operation
over. Defaults to 1, and reverts to 1 if x is not a list.
This argument may alternatively be a 'cluster' object,
in which case it is the user's responsibility to close the socket
connection at the conclusion of the operation,
for example by running \code{parallel::stopCluster(cores)}.
The string 'autodetect' is also accepted, in which case the maximum
number of cores to use is one less than the total number of cores available.}

\item{scores}{logical indicating whether the Akaike weight confidence
values should be attributed to the output object as a numeric vector
called "score". Defaults to TRUE.}

\item{paths}{logical indicating whether the path of each sequence
through the classification tree should be attributed to the output object
as a character vector called "path". Defaults to FALSE.}

\item{seqlengths}{logical indicating whether the original sequence
lengths should be attributed to the returned object.}

\item{threshs, minscores, minlengths, maxlengths}{logicals, indicating
which classification test results should be attributed to the output object
(for advanced use).}
}
\value{
a character string giving the lineage of the input sequence.
}
\description{
\code{"classify"} assigns a DNA barcode or other
  taxonomically-informative sequence to a node of a classification
  tree probabilistically, using a series of nested profile
  hidden Markov models.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\seealso{
\code{\link{learn}}
}
\author{
Shaun Wilkinson
}
