% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{contract}
\alias{contract}
\title{Contract a classification tree based on pattern matching.}
\usage{
contract(tree, clades = "", quiet = FALSE)
}
\arguments{
\item{tree}{an object of class \code{"insect"}.}

\item{clades}{a vector of character strings giving the binary indices
matching the labels of the nodes that are to be collapsed.}

\item{quiet}{logical indicating whether feedback should be printed
to the console.}
}
\value{
an object of class \code{"insect"}.
}
\description{
This function is used to collapse over-extended nodes.
}
\details{
TBA
}
\examples{
  ## TBA
}
\references{
TBA
}
\seealso{
\code{\link{expand}}
}
\author{
Shaun Wilkinson
}
