% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conversion}
\alias{conversion}
\alias{dna2char}
\alias{aa2char}
\alias{char2dna}
\alias{char2aa}
\title{Convert sequences between binary and string formats.}
\usage{
dna2char(x)

aa2char(x)

char2dna(z, simplify = FALSE)

char2aa(z, simplify = FALSE)
}
\arguments{
\item{x}{a "DNAbin" or "AAbin" object.}

\item{z}{a vector of concatenated strings representing DNA or amino acid sequences
(upper case).}

\item{simplify}{logical indicating whether length-one "DNAbin" or "AAbin" objects
should be simplified to vectors. Defaults to FALSE.}
}
\value{
\code{dna2char} and \code{aa2char} return vectors of concatenated strings
  (in upper case).
  char2dna and char2aa return "DNAbin" and "AAbin" objects, respectively.
}
\description{
These functions are used to convert DNA and amino acid sequences in
  "DNAbin" or "AAbin" format to character strings, and vice versa.
}
\details{
TBA
}
\examples{
##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
