% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguate.R
\name{disambiguate}
\alias{disambiguate}
\title{Convert oligonucleotide sequences to regular expressions.}
\usage{
disambiguate(z)
}
\arguments{
\item{z}{a concatenated string representing a DNA oligonucleotide sequence,
possibly with IUPAC ambiguity codes (all in upper case).}
}
\value{
a regular expression.
}
\description{
This function is used to convert an oligonucleotide sequence to a regular
  expression that can be used to query a sequence dataset and possibly
  remove primer sequences.
}
\details{
TBA
}
\examples{
disambiguate("GGWACWGGWTGAACWGTWTAYCCYCC")
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
