% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\name{encoding}
\alias{encoding}
\alias{encodePHMM}
\alias{decodePHMM}
\title{Encode and decode profile HMMs in raw bytes}
\usage{
encodePHMM(x)

decodePHMM(z)
}
\arguments{
\item{x}{an object of class "PHMM"}

\item{z}{a raw vector in the encodePHMM schema}
}
\value{
encodePHMM returns a raw vector. \code{decodePHMM} returns
  an object of class "PHMM".
}
\description{
These functions are used to compress and decompress profile
  hidden Markov models for DNA to improve memory efficiency.
}
\details{
TBA
}
\examples{
#TBA
}
\author{
Shaun Wilkinson
}
