% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork.R
\name{fork}
\alias{fork}
\title{Split nodes of a classification tree.}
\usage{
fork(node, x, refine = "Viterbi", iterations = 50, minK = 2, maxK = 2,
  minscore = 0.9, probs = 0.05, resize = TRUE, seqweights = "Gerstein",
  quiet = FALSE, ...)
}
\arguments{
\item{node}{an object of class \code{"insect"}.}

\item{x}{an object of class \code{"DNAbin"}.}

\item{refine}{character string giving the iterative model refinement
method to be used in the partitioning process. Valid options are
\code{"Viterbi"} (Viterbi training; the default option) and
\code{"BaumWelch"} (a modified version of the Expectation-Maximization
algorithm).}

\item{iterations}{integer giving the maximum number of training-classification
iterations to be used in the splitting process (see details section).
Note that this is not necessarily the same as the number of Viterbi training
or Baum Welch iterations to be used in model training, which can be set
using the argument \code{"maxiter"} (eventually passed to
\code{\link[aphid]{train}} via the dots argument "...").}

\item{minK}{integer. The minimum number of furications allowed at each inner
node of the tree. Defaults to 2 (all inner nodes are bifuricating).}

\item{maxK}{integer. The maximum number of furications allowed at each inner
node of the tree. Defaults to 2 (all inner nodes are bifuricating).}

\item{minscore}{numeric between 0 and 1. The minimum acceptable value
for the \emph{n}th percentile of Akaike weights (where \emph{n} is
the value given in \code{"probs"}, for the node to be split and the
recursion process to continue.
At any given node, if the \emph{n}th percentile of Akaike weights
falls below this threshold, the recursion process for the node will
terminate. As an extreme example, if \code{minscore = 0.9} and
\code{probs = 0.05} (the default settings), and after generating two
candidate PHMMs to occupy the candidate subnodes the lower 5th percentile
of Akaike weights is 0.89, the splitting process will
terminate and the function will simply return the unsplit root node.}

\item{probs}{numeric between 0 and 1. The percentile of Akaike weights
to test against the minimum score threshold given in \code{"minscore"}.}

\item{resize}{logical indicating whether the models should be free to
change size during the training process or if the number of modules
should be fixed. Defaults to TRUE. Only applicable if
\code{refine = "Viterbi"}.}

\item{seqweights}{an optional numeric vector the same length as the
sequence list giving the weights to use when training the
PHMMs. Alternatively its length can be the same as the total number
of unique sequences as specified in \code{duplicates}.
The character string "Gerstein" is also valid, which calls
the \code{\link[aphid]{weight}} function in the
\code{\link[aphid]{aphid}} package to derive the weights using the
algorithm of Gerstein et al. 1994.
If NULL, all sequences are given weights of 1.}

\item{quiet}{logical indicating whether feedback should be printed
to the console. Note that the output can be rather verbose for
larger trees.}

\item{...}{further arguments to be passed to \code{\link{partition}}
(which may then be passed on to \code{\link[aphid]{train}}).}
}
\value{
an object of class \code{"insect"}.
}
\description{
This function is used to split a leaf node of a classification tree,
  given a set of sequences.
}
\details{
Note that seqweights argument should have the same length as x.
}
\examples{
  ## TBA
}
\references{
TBA
}
\seealso{
\code{\link{learn}} (parent function),
  \code{\link{partition}} (child function).
}
\author{
Shaun Wilkinson
}
