% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{get_taxon}
\alias{get_taxon}
\title{Derive taxon ID from lineage string.}
\usage{
get_taxon(lineage, db)
}
\arguments{
\item{lineage}{character. A semicolon-delimited lineage string.}

\item{db}{taxonomic database in matrix form.}
}
\value{
an integer. The unique taxon database ID.
}
\description{
This function returns the unique taxon ID associated with a given
  semicolon-delimited lineage string.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
