% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{join}
\alias{join}
\title{Concatenate DNAbin objects while preserving attributes.}
\usage{
join(...)
}
\arguments{
\item{...}{\code{DNAbin} objects to be concatenated.}
}
\value{
an object of class \code{DNAbin}.
}
\description{
This function joins two or more \code{DNAbin} objects, as well as the
  attributes whose lengths match the objects themselves
}
\details{
TBA
}
\examples{
  ## TBA
}
\seealso{
\code{\link{subset.DNAbin}}
}
\author{
Shaun Wilkinson
}
