% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate.R
\name{manipulate}
\alias{manipulate}
\alias{duplicated.DNAbin}
\alias{unique.DNAbin}
\alias{subset.DNAbin}
\alias{duplicated.AAbin}
\alias{unique.AAbin}
\alias{subset.AAbin}
\title{Further bit-level manipulation of sequences.}
\usage{
\method{duplicated}{DNAbin}(x, incomparables = FALSE, pointers = TRUE, ...)

\method{unique}{DNAbin}(x, incomparables = FALSE, attrs = TRUE,
  drop = FALSE, ...)

\method{subset}{DNAbin}(x, subset, attrs = TRUE, drop = FALSE, ...)

\method{duplicated}{AAbin}(x, incomparables = FALSE, pointers = TRUE, ...)

\method{unique}{AAbin}(x, incomparables = FALSE, attrs = TRUE,
  drop = FALSE, ...)

\method{subset}{AAbin}(x, subset, attrs = TRUE, drop = FALSE, ...)
}
\arguments{
\item{x}{a \code{DNAbin} or \code{AAbin} object.}

\item{incomparables}{placeholder, not currently functional.}

\item{pointers}{logical indicating whether the re-replication key
should be returned as a \code{"pointers"} attribute of the output vector
(only applicable for \code{duplicated.DNAbin} and \code{duplicated.AAbin}).
Note that this can increase the computational
time for larger sequence lists.}

\item{...}{further agruments to be passed between methods.}

\item{attrs}{logical indicating whether the attributes of the input object
whose length match the object length (or number of rows if it is a matrix)
should be retained and subsetted along with the object.
This is useful if the input object has species, lineage and/or taxon ID
metadata that need to be retained following the duplicate analysis.}

\item{drop}{logical; indicates whether the input matrix (assuming one is
passed) should be reduced to a vector if subset to a single sequence.
Defaults to FALSE in keeping with the style of the \code{\link{ape}} package.}

\item{subset}{logical vector giving the elements or rows to be kept.}
}
\value{
\code{unique} and \code{subset} return a
  \code{DNAbin} or \code{AAbin} object. \code{duplicated} returns a logical vector.
}
\description{
These functions provide additional methods to manipulate objects of class
  \code{"DNAbin"}.
}
\details{
\code{duplicated} returns a logical vector indicating the sequences
  that are duplicated, and optionally provides a "pointers" attribute
  indicating which of the unique sequences they are duplicates of.
\code{unique} takes a list or matrix of DNA or amino acid sequences and returns
the subset of sequences that are unique.
}
\examples{
  ## find duplicates in a small subsection of the woodmouse alignment
  library(ape)
  data(woodmouse)
  x <- woodmouse[, 100:150]
  duplicates <- duplicated.DNAbin(x, point = TRUE)
  duplicates
  ## shows that there are 3 unique sequences, No305, No304, and No0910S
  ## pointers attribute shows that most are identical to the second one (No304)
  ##
  ## reduce the alignment to only include the unique sequences
  x <- unique.DNAbin(x)
  x
  ## further subset the alignment to only include the first 2 sequences
  x <- subset.DNAbin(x, subset = c(TRUE, TRUE, FALSE))
}
\references{
Paradis E (2007) A bit-level coding scheme for nucleotides. \url{
  http://ape-package.ird.fr/misc/BitLevelCodingScheme.html}.
}
\seealso{
\code{\link[ape]{DNAbin}}
}
\author{
Shaun Wilkinson
}
