% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{prune_taxon}
\alias{prune_taxon}
\title{Purge clades from taxon database.}
\usage{
prune_taxon(db, clade, keep = FALSE)
}
\arguments{
\item{db}{the taxon database, obtained by running
download_taxon.}

\item{clade}{the name or taxon ID number of the clade to be removed
(or kept if keep is TRUE).}

\item{keep}{logical, indicates whether the specified clade should be
kept, and the rest of the database purged. Defaults to FALSE.}
}
\value{
a smaller or equal sized dataframe with the same column names as
  the input object
  ("tax_id", "parent_tax_id", "rank", "name", "parent_tax_index").
}
\description{
This function prunes the taxon database, removing specified clades as
  desired to improve speed and memory efficiency.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
