% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{purge}
\alias{purge}
\title{Find and collapse over-extended nodes}
\usage{
purge(tree, quiet = FALSE)
}
\arguments{
\item{tree}{an object of class \code{"insect"}.}

\item{quiet}{logical indicating whether feedback should be printed
to the console.}
}
\value{
an object of class \code{"insect"}.
}
\description{
This function tests each node of a classification tree
  for over-extension by checking if all sequences belonging
  to the node have identical lineage metadata. Over-extended
  nodes are collapsed and the simplified tree returned.
}
\details{
TBA
}
\examples{
  ## TBA
}
\references{
TBA
}
\seealso{
\code{\link{expand}}, \code{\link{contract}}
}
\author{
Shaun Wilkinson
}
