% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{qc}
\alias{qc}
\title{Sequence quality filter.}
\usage{
qc(x, minqual = 30, maxambigs = 0, mincount = 2, minlength = 50,
  maxlength = 500)
}
\arguments{
\item{x}{a vector of concatenated strings representing DNA sequences
(in upper case) or a DNAbin list object with quality attributes.
This argument will usually be the output from \code{readFASTQ}.}

\item{minqual}{integer, the minimum average quality score for a
sequence to pass the filter. Defaults to 30.}

\item{maxambigs}{integer, the maximum number of ambiguities for a sequence
to pass the filter. Defaults to 0.}

\item{mincount}{integer, the minimum acceptable number of occurences of a
sequence for it to pass the filter. Defaults to 2 (removes singletons).}

\item{minlength}{integer, the minimum acceptable sequence length.
Defaults to 50.}

\item{maxlength}{integer, the maximum acceptable sequence length.
Defaults to 500.}
}
\value{
a possibly shortened object of the same type as the primary
  input argument
  (i.e. a "DNAbin" object if x is a "DNAbin" object and a vector
  of concatenated character strings otherwise).
}
\description{
This function checks that the average Phred quality scores of sequences (imported
  from a FASTQ file) are above a given threshold, and that the maximum number
  of ambiguities is below another given threshold.
}
\details{
TBA
}
\examples{
##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
