% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.R
\name{rc}
\alias{rc}
\title{Reverse complement DNA in character string format.}
\usage{
rc(z)
}
\arguments{
\item{z}{a vector of concatenated strings representing DNA sequences
(in upper case).}
}
\value{
a vector of concatenated strings representing DNA sequences
  (in upper case).
}
\description{
This function reverse complements a DNA sequence or set of DNA
  sequences that are in concatenated character string format.
}
\details{
TBA
}
\examples{
rc("TATTG")
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
