% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{readGB}
\alias{readGB}
\title{Read DNA sequences from GenBank.}
\usage{
readGB(accs, prompt = FALSE, contact = NULL, quiet = FALSE)
}
\arguments{
\item{accs}{character vector of GenBank accession numbers.}

\item{prompt}{logical indicating whether the user should be prompted with
the total number of sequences to download before commencing.}

\item{contact}{character string, giving users the option to provide an
emil address to send on the web form.}

\item{quiet}{logical indicating whether feedback should be printed
to the console.}
}
\value{
a list of DNA sequences as a \code{DNAbin} object.
}
\description{
This function downloads DNA sequence data from GenBank via the Entrez API.
}
\details{
TBA
}
\examples{
  ## TBA
}
\references{
TBA
}
\seealso{
\code{\link{searchGB}}
}
\author{
Shaun Wilkinson
}
