% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{replicate}
\alias{replicate}
\alias{dereplicate}
\alias{rereplicate}
\title{Dereplicate and rereplicate sequence datasets.}
\usage{
dereplicate(x, cores = 1)

rereplicate(x)
}
\arguments{
\item{x}{a list of sequences in \code{DNAbin} or \code{AAbin} format, or a
vector of sequences as concatenated upper-case character strings.}

\item{cores}{integer, the number of processors to use (advanced).}
}
\value{
either a DNAbin/AAbin object, or a vector of concatenated
  upper-case character strings, depending on the original input object.
}
\description{
These functions are used to extract only the unique sequences from
  a set of DNA reads, with the ability to rebuild the original
  sequence set.
}
\details{
TBA
}
\examples{
##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
