% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchBOLD}
\alias{searchBOLD}
\title{Query the BOLD database and download matching sequences.}
\usage{
searchBOLD(taxon, markers = c("COI-5P", "COI-3P", "28S", "16S", "COII",
  "CYTB", "atp6", "COXIII", "18S", "ITS", "ITS2", "ITS1", "5.8S"))
}
\arguments{
\item{taxon}{a recognized scientific taxon name.}

\item{markers}{character string or vector giving the genetic marker(s)
to limit the result to. Can be any or all of: "COI-5P", "COI-3P", "28S",
"16S", "COII", "CYTB", "atp6", "COXIII", "18S", "ITS", "ITS2", "ITS1", "5.8S".}
}
\value{
a list of DNA sequences as a \code{DNAbin} object.
}
\description{
This function searches the barcode of life (BOLD) database using the public
  API, and downloads the matching sequences as a \code{"DNAbin"} object.
  Note that an internet connection is required when using this function.
  In some cases depending on the user security settings the R session
  may need to be opened with administrator privileges.
}
\details{
This function calls the BOLD API
  (\url{http://www.boldsystems.org/index.php/resources/api?type=webservices})
  to search and download DNA barcode sequences from the BOLD database
  (Ratnasingham & Hebert 2007).
}
\examples{
  ## Query the BOLD database for Saccharomycetes ITS2 sequences
  \dontrun{
    x <- searchBOLD("Saccharomycetes", markers = "ITS2")
  }
}
\references{
Ratnasingham S & Hebert PDN (2007) BOLD: the barcode of life data
  system (www.barcodinglife.org). \emph{Molecular Ecology Notes},
  \strong{7}, 355–364.
}
\seealso{
\code{\link{searchGB}} for downloading DNA sequences from GenBank.
}
\author{
Shaun Wilkinson
}
