% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchGB}
\alias{searchGB}
\title{Query GenBank and download matching sequences.}
\usage{
searchGB(query, onlyID = FALSE, DNA = TRUE, prompt = TRUE,
  contact = NULL, quiet = FALSE)
}
\arguments{
\item{query}{an Entrez query provided as a character string. For help
compiling Entrez queries see
\url{https://www.ncbi.nlm.nih.gov/books/NBK3837/#EntrezHelp.Entrez_Searching_Options}
and \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/}.}

\item{onlyID}{logical. Should just the GenBank accession numbers be
returned (FALSE) or should the function return the entire sequences
and their metadata (TRUE; default)?}

\item{DNA}{logical indicating whether the returned value should be of
class \code{DNAbin} (TRUE; defult). If \code{FALSE} a list of character
vectors is returned. \code{DNA = TRUE} is recommended for memory and speed
efficiency and compatibility with other functions in the \code{insect}
package.}

\item{prompt}{logical indicating whether to check with the user before
downloading sequences.}

\item{contact}{an optional character string with the users email address.
This is added to the E-utilities URL and may be used by NCBI to contact
the user if the application causes unintended issues.}

\item{quiet}{logical indicating whether the progress should be printed
to the console.}
}
\value{
a list of DNA sequences, either in \code{DNAbin} format or as
  character vectors if \code{DNA} is set to \code{FALSE}.
}
\description{
\code{searchGB} searches the NCBI GenBank database using the Entrez
  utilities, and downloads the matching sequences either as a
  \code{"DNAbin"} object or a list of character vectors. Note
  that an internet connection is required when using this function.
  In some cases depending on the user security settings the R session
  may need to be opened with administrator privileges.
}
\details{
TBA
}
\examples{
  ## Query the GenBank database for Eukaryote mitochondrial 16S sequences
  ## between 100 and 300 base pairs in length and last modified between
  ## 1999 and 2000.
  \dontrun{
    query <- "Eukaryota[ORGN]+AND+16S[TITL]+AND+100:300[SLEN]+AND+1999:2000[MDAT]"
    x <- searchGB(query)
  }
}
\references{
TBA
}
\seealso{
\code{\link{readGB}} and \code{\link[ape]{read.GenBank}} (ape)
  for downloading DNA sequences from GenBank using accession numbers.
}
\author{
Shaun Wilkinson
}
