% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{sift}
\alias{sift}
\title{Identify sequences with potentially incorrect lineage metadata}
\usage{
sift(x, db, level = "order", threshold = 0.97, quiet = FALSE)
}
\arguments{
\item{x}{a DNAbin object with lineage attributes.}

\item{db}{the taxon database.}

\item{level}{taxonomic level at which heterogeneity within an OTUs flags 
a sequence as potentially erroneous. Must be a recognized rank
in the NCBI Taxon database.}

\item{threshold}{numeric between 0 and 1 giving the OTU similarity cutoff value.}

\item{quiet}{logical indicating whether progress should be printed to the console.}
}
\value{
a data frame containing the names of the potentially erroneous sequences.
}
\description{
This function evaluates a list of DNA barcode
   sequences (a "DNAbin" object with lineage attributes)
   and returns a table of sequences that
   may require further checking before being used as training
   data for downstream tree-learning operations.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
