% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Paired-end rread stitching.}
\usage{
stitch(fw, rv, mindiff = 6, cores = 1, ...)
}
\arguments{
\item{fw, rv}{DNAbin objects with quality attributes (see \code{\link{readFASTQ}} to
generate these objects from fastq text files), representing the forward and
reverse reads to be stitched. These objects can be either vectors or lists.
In the latter case, the two objects must be equal length.}

\item{mindiff}{the minimum difference in quality between two different base calls
for the higher quality call to be added to the consensus alignment. In cases where
the quality differences are less than this threshold, the ambiguity code "N" is added
to the consensus sequence.}

\item{cores}{The number of threads to be used to process the operation.}

\item{...}{further agruments to be passed to \code{\link[aphid]{align}}}
}
\value{
a DNAbin object with quality attributes
}
\description{
This function aligns forward and reverse reads generated from the Illumina
  2x amplicon sequencing platforms, and produces a consensus sequence with
  maximum Phred scores attached as "quality" attributes.
}
\details{
TBA
}
\examples{
  \dontrun{
    ##TBA
  }
}
\references{
TBA
}
\seealso{
\code{\link{readFASTQ}} for generating DNAbin objects from FASTQ text files
}
\author{
Shaun Wilkinson
}
