% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\alias{writeFASTQ}
\alias{writeFASTA}
\title{Write sequences to text in FASTA or FASTQ format.}
\usage{
writeFASTQ(x, file = "", append = FALSE)

writeFASTA(x, file = "", append = FALSE)
}
\arguments{
\item{x}{a list of sequences in \code{DNAbin} or \code{AAbin} format, or a
vector of sequences as character strings.
For writeFASTQ, only DNAbin objects are accepted, and each element should have
a vector of quality scores of equal length attributed to the sequence.
These vectors are comprised of raw bytes ranging from 00 to 5d
(0 to 93 when converted to integers; this format is more memory
efficient than the ASCII characters used in the FASTQ text files).
See \code{\link{readFASTQ}} for more details.}

\item{file}{character string giving a valid file path to output the text to.
If file = "" (default setting) the text file is written to the
console.}

\item{append}{logical indicating whether text should be appended below
existing text in the file (TRUE), or whether any existing text should be
overwritten (FALSE; default). Only applicable if the file specified
in \code{file} already exists.}
}
\description{
These functions take a list of DNA or amino acid sequences in
  \code{DNAbin} or \code{AAbin} format
  and outputs a text file to a specified directory.
}
\details{
TBA
}
\examples{
  \dontrun{
  ## TBA
  }
}
\references{
Illumina help page:
  \url{https://help.basespace.illumina.com/articles/descriptive/fastq-files/}
}
\seealso{
\code{\link{readFASTQ}} for reading FASTQ files into R,
  and \code{\link[ape]{write.dna}} in the ape package
  for writing DNA to text in FASTA and other formats.
}
\author{
Shaun Wilkinson
}
