% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeFASTQ}
\alias{writeFASTQ}
\title{Write DNA to text in FASTQ format.}
\usage{
writeFASTQ(x, file = "", append = FALSE)
}
\arguments{
\item{x}{a list object of class \code{DNAbin} with each element also having
a vector of quality scores of equal length to the sequence. These vectors
are comprised of raw bytes ranging from 00 to 5d (0 to 93 when converted to
integers; this format is more memory efficient than the ASCII characters used
in the FASTQ text files). See \code{\link{readFASTQ}} for more details.}

\item{file}{character string giving a valid file path to output the text to.
If file = "" (default setting) the text file is written to the
current working directory.}

\item{append}{logical indicating whether text should be appended below
existing text in the file (TRUE), or whether any existing text should be
overwritten (FALSE; default). Only applicable if the file specified
in \code{file} already exists.}
}
\description{
This function takes a list of DNA sequences in \code{DNAbin} format
  with "quality" attributes (e.g. Phred scores) and outputs
  a text file in FASTQ format to a specified directory.
}
\details{
TBA
}
\examples{
  \dontrun{
  ## TBA
  }
}
\references{
Illumina help page:
  \url{https://help.basespace.illumina.com/articles/descriptive/fastq-files/}
}
\seealso{
\code{\link{readFASTQ}} for reading FASTQ filed into R,
  and \code{\link[ape]{write.dna}} in the ape package
  for writing DNA to text in FASTA and other formats.
}
\author{
Shaun Wilkinson
}
