% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Identify sequences with potentially incorrect lineage metadata.}
\usage{
check(x, db, level = "order", threshold = 0.97, quiet = FALSE)
}
\arguments{
\item{x}{a DNAbin list object with "taxID" and/or "lineage" attributes
(see \code{\link{searchGB}} for details).}

\item{db}{a local copy of the NCBI Taxonomy database
(see \code{\link{download_taxon}} for details).}

\item{level}{character string giving the taxonomic level at which
heterogeneity within a cluster flags
a sequence as potentially erroneous. Must be a recognized rank
within the NCBI Taxonomy database.}

\item{threshold}{numeric between 0 and 1 giving the OTU similarity cutoff value
with which to cluster the sequences.}

\item{quiet}{logical indicating whether progress should be printed to the console.}
}
\value{
a data frame containing the names of the potentially erroneous sequences.
}
\description{
This function evaluates a list of DNA barcode
   sequences (a "DNAbin" object with lineage attributes)
   and returns a table of sequences that
   may require further checking before being used as training
   data for downstream tree-learning operations.
}
\details{
This function first clusters the sequence dataset into operational
  taxonomic units (OTUs) based on a given genetic distance threshold using the
  \code{\link[kmer]{otu}} function in the \code{\link{kmer}} package.
  It then proceeds to check each cluster
  for lineage homogeneity at a given taxonomic rank,
  flagging any records that appear out of place based on the taxa/lineages
  of the other OTU members.
}
\author{
Shaun Wilkinson
}
