% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Taxonomic classification for DNA barcode sequences.}
\usage{
classify(x, tree, threshold = 0.999, cores = 1)
}
\arguments{
\item{x}{a \code{"DNAbin"} object, either as a vector or a list of vectors.}

\item{tree}{an object of class \code{"insect"}.}

\item{threshold}{numeric value between 0 and 1 giving the minimum
Akaike weight for the recursive classification procedure
to continue toward the leaves of the tree.}

\item{cores}{integer giving the number of CPUs to parallelize the operation
over. Defaults to 1, and reverts to 1 if x is not a list.
This argument may alternatively be a 'cluster' object,
in which case it is the user's responsibility to close the socket
connection at the conclusion of the operation,
for example by running \code{parallel::stopCluster(cores)}.
The string 'autodetect' is also accepted, in which case the maximum
number of cores to use is one less than the total number of cores available.}
}
\value{
a character string giving the lineage of the input sequence
}
\description{
\code{"classify"} probabilistically assigns a DNA barcode or other
  taxonomically-informative sequence to a node of a classification
  tree using a series of nested profile hidden Markov models.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\seealso{
\code{\link{learn}}
}
\author{
Shaun Wilkinson
}
