% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Taxonomic classification for DNA barcode sequences.}
\usage{
classify(x, tree, threshold = 0.9)
}
\arguments{
\item{x}{a \code{"DNAbin"} object, either as a vector or a list of vectors.}

\item{tree}{an object of class \code{"insect"}.}

\item{threshold}{numeric value between 0 and 1 giving the minimum
Akaike weight for the recursive classification procedure
to continue toward the leaves of the tree.}
}
\value{
a character string giving the lineage of the input sequence
}
\description{
\code{"classify"} probabilistically assigns a DNA barcode or other
  taxonomically-informative sequence to a node of a classification
  tree using a series of nested profile hidden Markov models.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\seealso{
\code{\link{learn}}
}
\author{
Shaun Wilkinson
}
