% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguate.R
\name{disambiguate}
\alias{disambiguate}
\title{Convert oligonucleotide sequences into regular expressions.}
\usage{
disambiguate(z)
}
\arguments{
\item{z}{a concatenated string representing a DNA oligonucleotide sequence,
possibly with IUPAC ambiguity codes (all in upper case).}
}
\value{
a regular expression.
}
\description{
This function is used to convert an oligonucleotide sequence into a regular
  expression that can be used to query a sequence dataset.
  This is generally used for finding and removing primer,
  adapter and/or index sequences.
}
\examples{
disambiguate("GGWACWGGWTGAACWGTWTAYCCYCC")
}
\author{
Shaun Wilkinson
}
