% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{download_taxon}
\alias{download_taxon}
\title{Download taxon database}
\usage{
download_taxon()
}
\value{
a dataframe with the following columns: "tax_id", "parent_tax_id",
  "rank", "name", "parent_tax_index"
}
\description{
This function accesses the NCBI API and gets the latest copy of the taxon
  database
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
