% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork.R
\name{fork}
\alias{fork}
\title{Split nodes of a classification tree.}
\usage{
fork(node, x, lineages, refine = "Viterbi", nstart = 10, iterations = 50,
  minK = 2, maxK = 2, minscore = 0.9, probs = 0.05, retry = TRUE,
  resize = TRUE, maxsize = NULL, kmers = NULL, seqweights = "Gerstein",
  cores = 1, quiet = FALSE, ...)
}
\arguments{
\item{node}{an object of class \code{"insect"}.}

\item{x}{an object of class \code{"DNAbin"}.}

\item{lineages}{a character vector (the same length as x) of
semi-colon delimited strings providing the lineage metadata
for each sequence.}

\item{refine}{character string giving the iterative model refinement
method to be used in the partitioning process. Valid options are
\code{"Viterbi"} (Viterbi training; the default option) and
\code{"BaumWelch"} (a modified version of the Expectation-Maximization
algorithm).}

\item{nstart}{integer. The number of random starting sets to be chosen
for initial k-means assignment of sequences to groups.}

\item{iterations}{integer giving the maximum number of training-classification
iterations to be used in the splitting process (see details section).
Note that this is not necessarily the same as the number of Viterbi training
or Baum Welch iterations to be used in model training, which can be set
using the argument \code{"maxiter"} (eventually passed to
\code{\link[aphid]{train}} via the dots argument "...").}

\item{minK}{integer. The minimum number of furications allowed at each inner
node of the tree. Defaults to 2 (all inner nodes are bifuricating).}

\item{maxK}{integer. The maximum number of furications allowed at each inner
node of the tree. Defaults to 2 (all inner nodes are bifuricating).}

\item{minscore}{numeric between 0 and 1. The minimum acceptable value
for the \emph{n}th percentile of Akaike weights (where \emph{n} is
the value given in \code{"probs"}, for the node to be split and the
recursion process to continue.
At any given node, if the \emph{n}th percentile of Akaike weights
falls below this threshold, the recursion process for the node will
terminate. As an extreme example, if \code{minscore = 0.9} and
\code{probs = 0.05} (the default settings), and after generating two
candidate PHMMs to occupy the candidate subnodes the lower 5th percentile
of Akaike weights is 0.89, the splitting process will
terminate and the function will simply return the unsplit root node.}

\item{probs}{numeric between 0 and 1. The percentile of Akaike weights
to test against the minimum score threshold given in \code{"minscore"}.}

\item{retry}{logical indicating whether failure to split a node based on
the criteria outlined in 'minscore' and 'probs' should prompt a second
attempt with different initial groupings. These groupings are based on
maximum kmer frequencies rather than k-means division, which can give
suboptimal groupings when the cluster sizes are different (due to
the up-weighting of larger clusters in the k-means algorithm).}

\item{resize}{logical indicating whether the models should be free to
change size during the training process or if the number of modules
should be fixed. Defaults to TRUE. Only applicable if
\code{refine = "Viterbi"}.}

\item{maxsize}{integer giving the upper bound on the number of modules
in the PHMMs. If NULL (default) no maximum size is enforced.}

\item{kmers}{an optional matrix of k-mer frequencies used for
initial assignment of sequences to groups via k-means clustering.
Rows should sum to one to account for differences in sequence length.
Defaults to NULL, in which case k-mers are automatically
counted (using k = 5) and normalized to sequence length.}

\item{seqweights}{either NULL (all sequences are given weights
of 1), a numeric vector the same length as \code{x} representing
the sequence weights used to derive the model, or a character string giving
the method to derive the weights from the sequences. Currently only the
\code{"Gerstein"} method is supported (default). For this method, a
tree is first created by k-mer counting (see \code{\link[phylogram]{topdown}}),
and sequence weights are then derived from the tree using the 'bottom up'
algorithm of Gerstein et al (1994).}

\item{cores}{integer giving the number of CPUs to use
when training the models (only applicable if
\code{refine = 'Viterbi'}). Defaults to 1.
This argument may alternatively be a 'cluster' object,
in which case it is the user's responsibility to close the socket
connection at the conclusion of the operation,
e.g. by running \code{parallel::stopCluster(cores)}.
The string 'autodetect' is also accepted, in which case the maximum
number of cores to use is one less than the total number of cores
available.}

\item{quiet}{logical indicating whether feedback should be printed
to the console. Note that the output can be rather verbose for
larger trees.}

\item{...}{further arguments to be passed to \code{\link{partition}}
(which may then be passed on to \code{\link[aphid]{train}}).}
}
\value{
an object of class \code{"insect"}.
}
\description{
This function is used to split a leaf node of a classification tree,
  given a set of sequences.
}
\details{
Note that seqweights argument should have the same length as x.
}
\examples{
  ## TBA
}
\references{
Blackshields G, Sievers F, Shi W, Wilm A, Higgins DG (2010) Sequence embedding
  for fast construction of guide trees for multiple sequence alignment.
  \emph{Algorithms for Molecular Biology}, \strong{5}, 21.
}
\seealso{
\code{\link{learn}} (parent function),
  \code{\link{partition}} (child function).
}
\author{
Shaun Wilkinson
}
