% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{get_lineage}
\alias{get_lineage}
\title{Convert taxon ID to semicolon-delimited lineage string.}
\usage{
get_lineage(taxID, db)
}
\arguments{
\item{taxID}{integer, the taxon ID number.}

\item{db}{the NCBI taxon database (as a data.frame object).
See download_taxon for details.}
}
\value{
the full lineage as a semicolon-delimited lineage string.
}
\description{
This function returns the full lineage of a taxon ID number
  using the NCBI Taxon database.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
