% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{get_lineage}
\alias{get_lineage}
\title{Derive lineage from taxon ID.}
\usage{
get_lineage(id, db)
}
\arguments{
\item{id}{integer giving the taxon ID number}

\item{db}{taxonomic database in matrix form}
}
\value{
the full lineage as a named character vector of taxonomic ranks
}
\description{
This function returns the full lineage of a taxon ID number,
  given a taxonomic database.
}
\details{
TBA
}
\examples{
  ##TBA

}
\references{
TBA
}
\author{
Shaun Wilkinson
}
