% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon.R
\name{get_taxID}
\alias{get_taxID}
\title{Derive taxon ID from a given lineage.}
\usage{
get_taxID(lineage, db)
}
\arguments{
\item{lineage}{A semicolon-delimited lineage string or lineage name.}

\item{db}{the NCBI taxon database (as a data.frame object).
See download_taxon for details.}
}
\value{
The unique taxon database ID (integer).
}
\description{
This function returns the unique taxon ID associated with a given
  semicolon-delimited lineage string or lineage name,
  using the NCBI taxon database.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
