% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insect.R
\docType{package}
\name{insect}
\alias{insect}
\title{Informatic sequence classification trees.}
\description{
R package 'insect' contains functions to create and use
  classification trees for DNA meta-barcoding. This enables
  users to quickly and accurately assign the taxonomic identities of
  large numbers of DNA barcodes or other informative sequences
  generated on one of the standard high-throughput sequencing platforms.
}
