% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{join}
\alias{join}
\title{Concatenate DNAbin objects while preserving attributes.}
\usage{
join(...)
}
\arguments{
\item{...}{\code{DNAbin} objects to be concatenated.}
}
\value{
an object of class \code{DNAbin}.
}
\description{
This function joins two or more \code{DNAbin} objects, retaining any
  attributes whose lengths match those of the input objects (e.g. "species",
  "lineage" and/or "taxID" attributes).
}
\examples{
data(whales)
join(whales, whales)
}
\seealso{
\code{\link{subset.DNAbin}}.
}
\author{
Shaun Wilkinson
}
