% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Split sequences into subsets with maximum differentiation.}
\usage{
partition(x, model = NULL, K = 2, allocation = "cluster",
  refine = "Viterbi", nstart = 20, iterations = 50, kmers = NULL,
  seqweights = "Gerstein", cores = 1, quiet = FALSE, ...)
}
\arguments{
\item{x}{a \code{"DNAbin"} object containing multiple sequences in a list.}

\item{model}{an optional profile hidden Markov model as the hypothetical
data generating mechanism for the entire sequence set. This is used to
provide the starting parameters to train the top level models.}

\item{K}{integer. The number of groups to split the sequences into.}

\item{allocation}{integer vector giving the initial group membership of the
sequences. The vector length should be the same as that of the sequence set
\code{x}, and all elements should take values between 1 and K.
Alternatively, the character string "cluster" is accepted. In this case
sequences are initially allocated to K groups using K-means clustering.
This is the default option.}

\item{refine}{the method used to train/refine the models. Valid methods
are "Viterbi" (Viterbi training; default) and "BaumWelch" (a modified
version of the E-M algorithm).}

\item{nstart}{integer. The number of random starting sets to be chosen
for initial k-means assignment of sequences to groups.}

\item{iterations}{integer giving the maximum number of training-classification
iterations to be used in the splitting process (see details section).
Note that this is not necessarily the same as the number of Viterbi training
or Baum Welch iterations to be used in model training, which can be set
using the argument \code{"maxiter"} (eventually passed to
\code{\link[aphid]{train}} via the dots argument "...").}

\item{kmers}{an optional matrix of k-mer frequencies used for
initial assignment of sequences to groups via k-means clustering.
Rows should sum to one to account for differences in sequence length.
Defaults to NULL, in which case k-mers are automatically
counted (using k = 4) and normalized to sequence length.}

\item{seqweights}{either NULL (default; all sequences are given an equal
weight of 1), a numeric vector the same length as \code{x} representing
the sequence weights used to derive the model, or a character string giving
the method to derive the weights from the sequences. Currently only the
\code{"Gerstein"} method is supported (the default). For this method, a
tree is first created by k-mer counting (see the
\code{\link[kmer]{cluster}} function in the
\code{\link[kmer]{kmer}} package),
and sequence weights are derived from the tree using the 'bottom up'
algorithm of Gerstein et al. (1994). The sum of these weights are equal
to the number of sequences in the alignment (so that
\code{mean(seqweights) = 1}.
Note this does not need to be the case if providing weights as a numeric
vector).}

\item{cores}{integer giving the number of CPUs to use
when training the models (only applicable if
\code{refine = 'Viterbi'}). Defaults to 1.
This argument may alternatively be a 'cluster' object,
in which case it is the user's responsibility to close the socket
connection at the conclusion of the operation,
e.g. by running \code{parallel::stopCluster(cores)}.
The string 'autodetect' is also accepted, in which case the maximum
number of cores to use is one less than the total number of cores
available.}

\item{quiet}{logical indicating whether the progress should be printed to
the console.}

\item{...}{further arguments to be passed to \code{"aphid::train"} (not
including 'inserts' or 'cores').}
}
\value{
an object of class \code{split}. #############################TODO
}
\description{
Finds the best split of a set of DNA sequences and provides a profile
  hidden Markov model for each subset along with the logged full
  probability of each sequence given its respective model (calculated
  using the forward algorithm).
}
\details{
TBA.
}
\examples{
  ## TBA
}
\references{
Durbin R, Eddy SR, Krogh A, Mitchison G (1998) Biological
  sequence analysis: probabilistic models of proteins and nucleic acids.
  Cambridge University Press, Cambridge, United Kingdom.

  Gerstein M, Sonnhammer ELL, Chothia C (1994) Volume changes in protein evolution.
  \emph{Journal of Molecular Biology}, \strong{236}, 1067-1078.
}
\seealso{
\code{\link{fork}} (parent function)
}
\author{
Shaun Wilkinson
}
