% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{searchGB}
\alias{searchGB}
\title{Query GenBank and download matching sequences.}
\usage{
searchGB(query, onlyID = FALSE, prompt = TRUE, contact = NULL,
  quiet = FALSE)
}
\arguments{
\item{query}{an Entrez query provided as a character string. For help
compiling Entrez queries see
\url{https://www.ncbi.nlm.nih.gov/books/NBK3837/#EntrezHelp.Entrez_Searching_Options}
and \url{https://www.ncbi.nlm.nih.gov/books/NBK49540/}.}

\item{onlyID}{logical. Should just the GenBank accession numbers be
returned (FALSE) or should the function return the entire sequences
and their metadata (TRUE; default)?}

\item{prompt}{logical indicating whether to check with the user before
downloading sequences.}

\item{contact}{an optional character string with the users email address.
This is added to the E-utilities URL and may be used by NCBI to contact
the user if the application causes unintended issues.}

\item{quiet}{logical indicating whether the progress should be printed
to the console.}
}
\value{
a list of DNA sequences as a \code{DNAbin} object.
}
\description{
\code{searchGB} searches the NCBI GenBank database using the Entrez
  utilities, and downloads the matching sequences as a
  \code{"DNAbin"} object. Note
  that an internet connection is required when using this function.
  In some cases depending on the user security settings the R session
  may need to be opened with administrator privileges.
}
\details{
This function uses the Entrez e-utilities API to search and download
  sequences from GenBank.
  Occasionally users may encounter an unknown error that is not reproducible
  and appears to be related to database records being updated in GenBank.
  This can generally be remedied by re-running the function. If problems
  persist please feel free to raise an issue on the package bug-reports page at
  <http://github.com/shaunpwilkinson/insect/issues>.
}
\examples{
  ## Query the GenBank database for Eukaryote mitochondrial 16S sequences
  ## between 100 and 300 base pairs in length and last modified between
  ## 1999 and 2000.
  \dontrun{
    query <- "Eukaryota[ORGN]+AND+16S[TITL]+AND+100:300[SLEN]+AND+1999:2000[MDAT]"
    x <- searchGB(query)
  }
}
\references{
NCBI Resource Coordinators (2012) Database resources of the National
  Center for Biotechnology Information. \emph{Nucleic Acids Research},
   \strong{41} (Database issue): D8–D20.
}
\seealso{
\code{\link{readGB}} and \code{\link[ape]{read.GenBank}} (ape)
  for downloading DNA sequences from GenBank using accession numbers.
}
\author{
Shaun Wilkinson
}
