% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{sift}
\alias{sift}
\title{Identify sequences with potentially incorrect lineage metadata}
\usage{
sift(x, db, ranks = c("phylum", "class", "order"), threshold = 0.97)
}
\arguments{
\item{x}{a DNAbin object with lineage attributes.}

\item{db}{the taxon database.}

\item{ranks}{character vector of taxonomic ranks to compare, in order of priority.}

\item{threshold}{numeric between 0 and 1 giving the OTU similarity cutoff.}
}
\value{
a character vector giving the names of the potentially incorrect sequences
}
\description{
This function evaluates a list of DNA barcode
   sequences (a "DNAbin" object with lineage attributes)
   and returns a table of sequences that
   may require further checking before being used as training
   data for downstream tree-learning operations.
}
\details{
TBA
}
\examples{
  ##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
