% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Paired-end read stitching.}
\usage{
stitch(R1, R2, up = NULL, down = NULL, mindiff = 6, cores = 1)
}
\arguments{
\item{R1, R2}{DNAbin objects with quality attributes (see \code{\link{readFASTQ}} to
generate these objects from fastq text files), representing the forward and
reverse reads to be stitched. These objects can be either vectors or lists.
In the latter case, the two objects must be equal length.}

\item{up, down}{forward and reverse primer sequences (either as concatenated
character strings or "DNAbin" objects). Either both or neither should be provided.}

\item{mindiff}{the minimum difference in quality between two different base calls
for the higher quality call to be added to the consensus alignment. In cases where
the quality differences are less than this threshold, the ambiguity code "N" is added
to the consensus sequence.}

\item{cores}{The number of threads to be used to process the operation.}
}
\value{
a "DNAbin" object or a vector of concatenated character strings,
  depending on the input.
}
\description{
This function aligns forward and reverse reads generated from the Illumina
  2x amplicon sequencing platforms, and produces a consensus sequence with
  maximum Phred scores attached as "quality" attributes.
}
\details{
TBA
}
\examples{
  \dontrun{
    ##TBA
  }
}
\references{
TBA
}
\seealso{
\code{\link{readFASTQ}} for reading FASTQ text files into R.
}
\author{
Shaun Wilkinson
}
