% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulize.R
\name{tabulize}
\alias{tabulize}
\title{Tabulate classification output.}
\usage{
tabulize(y, db, aggregated = FALSE, ranks = c("kingdom", "phylum", "class",
  "order", "family", "genus", "species"))
}
\arguments{
\item{y}{a character vector of semicolon-delimited lineage strings
or a list of such vectors, as output from the function \code{classify}.}

\item{db}{a copy of the NCBI taxonomy database.
See \code{\link{download_taxon}} for details.}

\item{aggregated}{logical indicating whether the output table
should be aggregated by taxon ID, rather than having one row
for each unique sequence. Defaults to FALSE.}

\item{ranks}{character vector giving the taxonomic ranks to be
included in the output table. Must be a valid rank from the
NCBI taxonomy database (see https://www.ncbi.nlm.nih.gov/taxonomy).}
}
\value{
a data frame.
}
\description{
This function presents taxonomic classifications in tabular format
  for ease of interpretation, including sequence counts for each sample.
  This makes for easy exporting to csv or xlsx files.
}
\details{
TBA
}
\examples{
##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
