% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim primers.}
\usage{
trim(x, up, down = NULL)
}
\arguments{
\item{x}{a "DNAbin" object or a vector of concatenated upper-case character strings,
representing the sequences to be trimmed.}

\item{up, down}{"DNAbin" objects or  vectors of concatenated upper-case character strings,
representing the primer sequences.}
}
\value{
a "DNAbin" object or a vector of concatenated character strings,
  depending on the input.
}
\description{
This function trims the primer sequence(s) from a set of DNA sequences.
}
\details{
TBA
}
\examples{
##TBA
}
\references{
TBA
}
\author{
Shaun Wilkinson
}
