% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim}
\alias{trim}
\title{Trim ends of DNA sequences}
\usage{
trim(x, motif, direction = "both", cores = 1, ...)
}
\arguments{
\item{x}{an object of class \code{DNAbin}.}

\item{motif}{a \code{DNAbin} or \code{PHMM} object.}

\item{direction}{character string indicating
the direction of the trim. Options are "forward" (trims everything to
the right of the motif), "backward" (trims everything to the left of
the motif) or "both" (retains the motif region only).}

\item{cores}{integer giving the number of CPUs to parallelize the operation
over. Defaults to 1, and reverts to 1 if x is not a list.
This argument may alternatively be a 'cluster' object,
in which case it is the user's responsibility to close the socket
connection at the conclusion of the operation,
for example by running \code{parallel::stopCluster(cores)}.
The string 'autodetect' is also accepted, in which case the maximum
number of cores to use is one less than the total number of cores available.
Note that in this case there
may be a tradeoff in terms of speed depending on the number and size
of sequences to be processed, due to the extra time required to initialize
the cluster.}

\item{...}{further arguments to be passed to \code{\link[aphid]{Viterbi}}
(not including 'type').}
}
\value{
an object of class \code{DNAbin}.
}
\description{
This function uses the Viterbi algorithm to semi-globally align a motif to
  a DNA sequence, and removes all nucleotides to the left or right of the
  motif.
}
\details{
This functions finds the optimal semiglobal alignment (a.k.a. "glocal"
  alignment or global alignment with free end gaps) between a
  sequence \code{"x"} and a shorter sequence \code{"motif"}, returning
  the motif region of x along with the nucleotides to the left or right
  if \code{direction} is set to \code{"reverse"} or \code{"forward"},
  respectively.
}
\examples{
  ## TBA
}
\seealso{
\code{\link{virtualPCR}}
}
\author{
Shaun Wilkinson
}
